[Feat 3481] handled the managments of "soft links" on AutomatedTest

+++ tm/tm.service/src/main/java/org/squashtest/tm/service/internal/repository/AutomatedTestDao.java
@@ -54,7 +54,7 @@ public interface AutomatedTestDao {
 	 * @param testId
 	 * @return
 	 */
-	int countReferences(long testId);
+	long countReferences(long testId);
 
 
 	/**
